moved hero and updated imports
